Integrate Python root input regressions into existing input tests
